+++ VisualizationAPI/src/org/gephi/visualization/apiimpl/shape/Rectangle.java
@@ -79,10 +79,10 @@ class Rectangle extends AbstractShape {
 
     @Override
     protected boolean intersectsCircle(int x, int y, int radius) {
-        return (origin.x - x) * (origin.x - x) + (origin.y - y) * (origin.y - y) <= radius * radius ||
-               (origin.x - x) * (origin.x - x) + (opposite.y - y) * (opposite.y - y) <= radius * radius ||
-               (opposite.x - x) * (opposite.x - x) + (origin.y - y) * (origin.y - y) <= radius * radius ||
-               (opposite.x - x) * (opposite.x - x) + (opposite.y - y) * (opposite.y - y) <= radius * radius;
+        return (origin.x - x) * (origin.x - x) + (origin.y - y) * (origin.y - y) <= (float) radius * radius ||
+               (origin.x - x) * (origin.x - x) + (opposite.y - y) * (opposite.y - y) <= (float) radius * radius ||
+               (opposite.x - x) * (opposite.x - x) + (origin.y - y) * (origin.y - y) <= (float) radius * radius ||
+               (opposite.x - x) * (opposite.x - x) + (opposite.y - y) * (opposite.y - y) <= (float) radius * radius;
     }
 
     public SelectionType getSelectionType() {
